+++ pytripgui/canvas_vc/objects/data_base.py
@@ -3,15 +3,17 @@ from typing import Optional
 
 from pytrip import Cube
 
+from pytripgui.canvas_vc.projection_selector import ProjectionSelector
+
 
 class PlotDataBase(ABC):
-    def __init__(self, selector):
+    def __init__(self, selector: ProjectionSelector):
         self.aspect = 1.0  # aspect ratio of plot
 
         self.cube: Optional[Cube] = None  # placeholder for cube object from pytrip
         self.data_to_plot = None  # placeholder for extracted and prepared data to plot
 
-        self.projection_selector = selector
+        self.projection_selector: ProjectionSelector = selector
 
     @abstractmethod
     def prepare_data_to_plot(self):
