+++ pytripgui/canvas_vc/plotter/images/ctx_image.py
@@ -20,19 +20,3 @@ class CtxImage(PatientImageBase):
                                         extent=extent,
                                         origin='lower',
                                         zorder=self.zorder)
-        self._set_labels(data)
-
-    def _set_labels(self, data: Ctx):
-        plane = data.projection_selector.plane
-        # "Transversal" (xy)
-        if plane == "Transversal":
-            self._axes.set_xlabel('x [mm]')
-            self._axes.set_ylabel('y [mm]')
-        # "Sagittal" (yz)
-        elif plane == "Sagittal":
-            self._axes.set_xlabel('y [mm]')
-            self._axes.set_ylabel('z [mm]')
-        # "Coronal"  (xz)
-        elif plane == "Coronal":
-            self._axes.set_xlabel('x [mm]')
-            self._axes.set_ylabel('z [mm]')
